Split into multiple go modules
