--- go.mod
@@ -1,30 +1,0 @@
-module github.com/pulumi/pulumi-eks
-
-go 1.15
-
-require (
-	github.com/blang/semver v3.5.1+incompatible
-	github.com/docker/docker v1.13.1 // indirect
-	github.com/googleapis/gnostic v0.2.0 // indirect
-	github.com/imdario/mergo v0.3.11 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.1 // indirect
-	github.com/onsi/ginkgo v1.12.0 // indirect
-	github.com/onsi/gomega v1.9.0 // indirect
-	github.com/pkg/errors v0.9.1
-	github.com/pulumi/pulumi-aws/sdk/v3 v3.25.1
-	github.com/pulumi/pulumi-kubernetes/sdk/v2 v2.7.7
-	github.com/pulumi/pulumi/pkg/v2 v2.18.2
-	github.com/pulumi/pulumi/sdk/v2 v2.18.2
-	github.com/stretchr/testify v1.6.1
-	gopkg.in/airbrake/gobrake.v2 v2.0.9 // indirect
-	gopkg.in/gemnasium/logrus-airbrake-hook.v2 v2.1.2 // indirect
-	gopkg.in/inf.v0 v0.9.1 // indirect
-	gopkg.in/yaml.v2 v2.3.0
-	k8s.io/api v0.0.0-20190313235455-40a48860b5ab
-	k8s.io/apimachinery v0.0.0-20190313205120-d7deff9243b1
-	k8s.io/client-go v11.0.0+incompatible
-	k8s.io/klog v1.0.0 // indirect
-	k8s.io/utils v0.0.0-20190308190857-21c4ce38f2a7 // indirect
-	sigs.k8s.io/yaml v1.1.0 // indirect
-)
