+++ provider/go.mod
@@ -1,0 +1,13 @@
+module github.com/pulumi/pulumi-eks/provider
+
+go 1.15
+
+require (
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.1 // indirect
+	github.com/pkg/errors v0.9.1
+	github.com/pulumi/pulumi/pkg/v2 v2.18.2
+	github.com/pulumi/pulumi/sdk/v2 v2.18.2
+	github.com/spf13/pflag v1.0.5 // indirect
+	gopkg.in/yaml.v2 v2.3.0 // indirect
+)
