+++ sdk/go.mod
@@ -1,0 +1,12 @@
+module github.com/pulumi/pulumi-eks/sdk
+
+go 1.15
+
+require (
+	github.com/blang/semver v3.5.1+incompatible
+	github.com/pkg/errors v0.9.1
+	github.com/pulumi/pulumi-aws/sdk/v3 v3.25.1
+	github.com/pulumi/pulumi-kubernetes/sdk/v2 v2.7.7
+	github.com/pulumi/pulumi/sdk/v2 v2.18.2
+	gopkg.in/yaml.v2 v2.3.0 // indirect
+)
