+++ sdk/go/eks/clusterCreationRoleProvider.go
