+++ sdk/go/eks/doc.go
