+++ sdk/go/eks/init.go
