+++ sdk/go/eks/managedNodeGroup.go
