+++ sdk/go/eks/nodeGroupSecurityGroup.go
