+++ sdk/go/eks/provider.go
