+++ sdk/go/eks/pulumiUtilities.go
