+++ sdk/go/eks/vpcCni.go
